deep_agents_reports/pyproject.toml
[project]
name = "deep-agents-reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.8.0",
    "deepagents>=0.2.0",
    "fastmcp>=2.10.5",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-mcp>=0.1.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.35",
    "langgraph>=0.2.0",
    "mcp[cli]>=1.16.0",
    "neo4j>=5.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tiktoken>=0.11.0",
    "typing-extensions>=4.0.0",
]
